Add homepage utility functions for handling class notes and assignments snapshots
